refactor(app): split app into components
