+++ client/src/app.rs
@@ -31,4 +31,6 @@ mod plots;
 mod tab;
 mod tables;
 mod types;
+mod ui_components;
 mod utils;
+
