+++ client/src/app/ui_components.rs
@@ -1,0 +1,1 @@
+pub mod types;
